Improve scripts; Update tsconfig.json
