+++ api/src/db.ts
@@ -11,13 +11,13 @@ const prismaDb = new PrismaClient();
  * ```
  */
 export function exclude<T, Key extends keyof T>(
-    obj: T,
-    ...keys: Key[]
+	obj: T,
+	...keys: Key[]
 ): Omit<T, Key> {
-    for (const key of keys) {
-        delete obj[key];
-    }
-    return obj;
+	for (const key of keys) {
+		delete obj[key];
+	}
+	return obj;
 }
 
 export default prismaDb;
