+++ extension/.eslintrc.js
@@ -2,5 +2,5 @@
  * @type {import("eslint").Linter.Config}
  */
 module.exports = {
-    extends: ["../.eslintrc.js"],
+	extends: ["../.eslintrc.js"],
 };
