+++ extension/vsc-extension-quickstart.md
