--- server/package.json
@@ -1,23 +1,0 @@
-{
-    "name": "codesync-server",
-    "private": "true",
-    "scripts": {
-        "build": "tsc --project ./",
-        "build:watch": "tsc --watch --project ./",
-        "db:gen": "prisma generate",
-        "db:push": "prisma db push",
-        "dev": "tsc --watch --project ./",
-        "lint": "eslint .",
-        "prettier": "prettier --write --loglevel warn ."
-    },
-    "dependencies": {
-        "@fastify/cors": "^8.1.0",
-        "@prisma/client": "^4.1.1",
-        "dotenv": "^16.0.1",
-        "fastify": "^4.3.0",
-        "fastify-plugin": "^4.0.0"
-    },
-    "devDependencies": {
-        "prisma": "^4.1.1"
-    }
-}
