+++ ui/.eslintrc.js
@@ -2,6 +2,6 @@
  * @type {import("eslint").Linter.Config}
  */
 module.exports = {
-    plugins: ["eslint-plugin-solid"],
-    extends: ["../.eslintrc.js", "plugin:solid/typescript"],
+	plugins: ["eslint-plugin-solid"],
+	extends: ["../.eslintrc.js", "plugin:solid/typescript"],
 };
