+++ ui/index.html
@@ -1,20 +1,20 @@
 <!DOCTYPE html>
 <html lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <meta name="theme-color" content="#000000" />
-        <link
-            rel="shortcut icon"
-            type="image/ico"
-            href="/src/assets/favicon.ico"
-        />
-        <title>Solid App</title>
-    </head>
-    <body>
-        <noscript>You need to enable JavaScript to run this app.</noscript>
-        <div id="root"></div>
+	<head>
+		<meta charset="utf-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1" />
+		<meta name="theme-color" content="#000000" />
+		<link
+			rel="shortcut icon"
+			type="image/ico"
+			href="/src/assets/favicon.ico"
+		/>
+		<title>Solid App</title>
+	</head>
+	<body>
+		<noscript>You need to enable JavaScript to run this app.</noscript>
+		<div id="root"></div>
 
-        <script src="/src/index.tsx" type="module"></script>
-    </body>
+		<script src="/src/index.tsx" type="module"></script>
+	</body>
 </html>
