+++ ui/src/global.d.ts
@@ -1,6 +1,6 @@
 import * as _vscode from "vscode";
 
 declare global {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const vscodeApi: any;
+	// eslint-disable-next-line @typescript-eslint/no-explicit-any
+	const vscodeApi: any;
 }
