+++ ui/tsconfig.json
@@ -1,17 +1,18 @@
 {
-    "extends": "../tsconfig.json",
-    "compilerOptions": {
-        "strict": true,
-        "target": "ESNext",
-        "module": "ESNext",
-        "moduleResolution": "node",
-        "allowSyntheticDefaultImports": true,
-        "esModuleInterop": true,
-        "lib": ["DOM"],
-        "jsx": "preserve",
-        "jsxImportSource": "solid-js",
-        "types": ["vite/client"],
-        "noEmit": true,
-        "isolatedModules": true
-    }
+	"compilerOptions": {
+		"strict": true,
+		"target": "ESNext",
+		"module": "ESNext",
+		"outDir": "dist",
+		"moduleResolution": "Node",
+		"allowSyntheticDefaultImports": true,
+		"esModuleInterop": true,
+		"rootDir": "src",
+		"lib": ["ESNext", "DOM"],
+		"jsx": "preserve",
+		"jsxImportSource": "solid-js",
+		"types": ["vite/client"],
+		"noEmit": true,
+		"isolatedModules": true
+	}
 }
