+++ ui/vite.config.ts
@@ -4,22 +4,22 @@ import Unocss from "unocss/vite";
 import presetUno from "@unocss/preset-uno";
 
 export default defineConfig({
-    plugins: [
-        solidPlugin(),
-        Unocss({
-            presets: [presetUno()],
-            theme: {
-                colors: {
-                    vsblue: "#0066b8",
-                    vsgreen: "#008000",
-                },
-            },
-        }),
-    ],
-    server: {
-        port: 3000,
-    },
-    build: {
-        target: "esnext",
-    },
+	plugins: [
+		solidPlugin(),
+		Unocss({
+			presets: [presetUno()],
+			theme: {
+				colors: {
+					vsblue: "#0066b8",
+					vsgreen: "#008000",
+				},
+			},
+		}),
+	],
+	server: {
+		port: 3000,
+	},
+	build: {
+		target: "esnext",
+	},
 });
