--- vscode/tsconfig.json
@@ -1,15 +1,0 @@
-{
-    "extends": "../tsconfig.json",
-    "compilerOptions": {
-        "module": "CommonJS",
-        "target": "ES2020",
-        "outDir": "dist",
-        "lib": ["ES2020", "DOM"],
-        "sourceMap": true,
-        "rootDir": "src",
-        "strict": true,
-        "noImplicitReturns": true,
-        "noFallthroughCasesInSwitch": true,
-        "noUnusedParameters": true
-    }
-}
